% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{add_demography}
\alias{add_demography}
\title{Add demography}
\usage{
add_demography(p, demography)
}
\arguments{
\item{p}{parameter list}

\item{demography}{site demography}
}
\value{
modified parameter list
}
\description{
Add demography
}
