% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_interventions}
\alias{add_interventions}
\title{Add all interventions}
\usage{
add_interventions(p, interventions, species)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}

\item{species}{Can be falciparum: "pf" or vivax: "pv", for vivax SMC, RTSS
and PMC are not implemented}
}
\value{
modified parameter list
}
\description{
Add all interventions
}
