% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_irs}
\alias{add_irs}
\title{Add IRS}
\usage{
add_irs(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add IRS
}
