% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_itns}
\alias{add_itns}
\title{Add ITNs}
\usage{
add_itns(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add ITNs
}
