% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_pev_epi}
\alias{add_pev_epi}
\title{Add pre-erythrocytic vaccine}
\usage{
add_pev_epi(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add pre-erythrocytic vaccine
}
