% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_pmc}
\alias{add_pmc}
\title{Add PMC}
\usage{
add_pmc(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add PMC
}
