% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{add_seasonality}
\alias{add_seasonality}
\title{Add seasonality}
\usage{
add_seasonality(p, seasonality)
}
\arguments{
\item{p}{parameter list}

\item{seasonality}{site seasonality}
}
\value{
modified parameter list
}
\description{
Add seasonality
}
