% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_smc}
\alias{add_smc}
\title{Add SMC}
\usage{
add_smc(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add SMC
}
