% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{add_vectors}
\alias{add_vectors}
\title{Add vectors}
\usage{
add_vectors(p, vectors)
}
\arguments{
\item{p}{parameter list}

\item{vectors}{site vectors}
}
\value{
modified parameter list
}
\description{
Add vectors
}
