% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin_demography}
\alias{burnin_demography}
\title{Expand demography retrospectively for burn in period}
\usage{
burnin_demography(demography, burnin)
}
\arguments{
\item{demography}{Site demography inputs}

\item{burnin}{Burn in period (years)}
}
\value{
Demography inputs with burn in
}
\description{
Assumes demography for burn in period is = to those in year 1.
}
