% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin_interventions}
\alias{burnin_interventions}
\title{Expand interventions retrospectively for burn in period}
\usage{
burnin_interventions(interventions, burnin)
}
\arguments{
\item{interventions}{Site intervention inputs}

\item{burnin}{Burn in period (years)}
}
\value{
Intervention inputs with burn in
}
\description{
Assumes interventions for burn in period are = to those in year 1.
}
