% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{set_age_outputs}
\alias{set_age_outputs}
\title{Set age outputs}
\usage{
set_age_outputs(p, min_ages)
}
\arguments{
\item{p}{parameter list}

\item{min_ages}{Minimum age bands for incidence and n_age outputs}
}
\value{
modified parameter list
}
\description{
Set age outputs
}
