% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{single_site}
\alias{single_site}
\title{Extract a single site-input from a country site file}
\usage{
single_site(site_file, index)
}
\arguments{
\item{site_file}{Country site file}

\item{index}{Index row from site_file$sites}
}
\value{
Single site
}
\description{
Extract a single site-input from a country site file
}
