% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-package.R
\docType{package}
\name{site-package}
\alias{site}
\alias{site-package}
\title{site: Site specific malariasimulation modelling}
\description{
This package converts a parameterised "site file" into malariasimulation model input parameter list.
}
\author{
\strong{Maintainer}: Pete Winskill \email{p.winskill@imperial.ac.uk}

}
\keyword{internal}
